+++ examples/prompts/colored-prompt.py
@@ -1,8 +1,6 @@
-from quo.color import Color
 from quo.prompt import Prompt
 
-style = Color("fg:red")
 
-session = Prompt(style=style)
+session = Prompt(fg="red")
 
 session.prompt(">> ")