+++ examples/prompts/terminal.py
@@ -1,0 +1,6 @@
+from quo.prompt import Prompt
+
+
+session = Prompt()
+
+session.prompt("<red>john</red><white>@</white><green>localhost</green><red>:</red><cyan><u>/user/john</u></cyan><purple>$ </purple>")